Move code into the actual class and out of main
I added a lot of changes to try and create more functions into the class and out of main.  This will make it more reusable if we had different students, we could just make different Transc objects for each student, instead of putting a ton of extra code in main for each student.